refactor(backend): remove misc service
